feat: cherry picks from release for `aggsender` (#246)
* refactor: encode metadata to be backwards compatible (#216)

* refactor: encode metadata to be backwards compatible

* tests: fix UT

* apply feedback

* test: fix

* fix merge issues

* refactor: metadata cert versioning

* test: fix ut

* fix: typo

* use the CertificateBuildParams

* test: fix ut

* Apply feedback

* fix: lint

* fix: Aggsender high cpu usage (#235)

- fix aggsender is using about 200% of CPU
- add a warning if decrease blockNumber (from L1 provider)

* fix: calculate correct root for `LER` merkle proof and fix `BridgeExit` hash (#237)

* fix: calculate correct root for LER merkle proof, and fix BridgeExit hash

* fix: UTs

* fix: comments

* fix: sqlite3 config improvements

Set some recommended params for sqlite connection.

* fix: lint

---------

+++ aggsender/db/aggsender_db_storage_test.go
@@ -29,7 +29,7 @@ func Test_Storage(t *testing.T) {
 	storage, err := NewAggSenderSQLStorage(log.WithFields("aggsender-db"), cfg)
 	require.NoError(t, err)
 
-	updateTime := time.Now().UTC().UnixMilli()
+	updateTime := uint32(time.Now().UTC().UnixMilli())
 
 	t.Run("SaveLastSentCertificate", func(t *testing.T) {
 		certificate := types.CertificateInfo{
@@ -242,7 +242,7 @@ func Test_SaveLastSentCertificate(t *testing.T) {
 	storage, err := NewAggSenderSQLStorage(log.WithFields("aggsender-db"), cfg)
 	require.NoError(t, err)
 
-	updateTime := time.Now().UTC().UnixMilli()
+	updateTime := uint32(time.Now().UTC().UnixMilli())
 
 	t.Run("SaveNewCertificate", func(t *testing.T) {
 		certificate := types.CertificateInfo{
